+++ src/pybind/py_e2sar.cpp
@@ -43,6 +43,7 @@ void init_e2sarCP(py::module_ &m);    // in a submodule "ControlPlane"
 void init_e2sarUtil(py::module_ &m);  // in the main module
 void init_e2sarAffinity(py::module_ &m);  // in the main module
 void init_e2sarHeaders(py::module_ &m);  // in the main module
+void init_e2sarNetUtil(py::module_ &m);  // in the main module
 void init_e2sarDP(pybind11::module_ &m);    // in a submodule "DataPlane"
 
 // "e2sar_py" will be the python module name using in "import xxx"
@@ -118,6 +119,7 @@ PYBIND11_MODULE(e2sar_py, m) {
     init_e2sarCP(m);
     init_e2sarDP(m);
     init_e2sarAffinity(m);
+    init_e2sarNetUtil(m);
 }